Redefine SQLTranslationException's vendor codes
