+++ shardingsphere-kernel/shardingsphere-sql-translator/shardingsphere-sql-translator-api/src/main/java/org/apache/shardingsphere/sqltranslator/exception/syntax/UnsupportedTranslatedDatabaseException.java
@@ -15,18 +15,20 @@
  * limitations under the License.
  */
 
-package org.apache.shardingsphere.sqltranslator.exception;
+package org.apache.shardingsphere.sqltranslator.exception.syntax;
 
+import org.apache.shardingsphere.infra.database.type.DatabaseType;
 import org.apache.shardingsphere.infra.util.exception.external.sql.sqlstate.XOpenSQLState;
+import org.apache.shardingsphere.sqltranslator.exception.SQLTranslationException;
 
 /**
- * Unsupported translated SQL exception.
+ * Unsupported translated database exception.
  */
-public final class UnsupportedTranslatedSQLException extends SQLTranslationException {
+public final class UnsupportedTranslatedDatabaseException extends SQLTranslationException {
     
-    private static final long serialVersionUID = -1419778194546662319L;
+    private static final long serialVersionUID = -8311552562051028033L;
     
-    public UnsupportedTranslatedSQLException(final String sql) {
-        super(XOpenSQLState.SYNTAX_ERROR, 1, "Translation error, SQL is: %s", sql);
+    public UnsupportedTranslatedDatabaseException(final DatabaseType databaseType) {
+        super(XOpenSQLState.SYNTAX_ERROR, 20, "Can not support database `%s` in SQL translation", databaseType.getType());
     }
 }
