+++ shardingsphere-kernel/shardingsphere-sql-translator/shardingsphere-sql-translator-core/src/test/java/org/apache/shardingsphere/sqltranslator/rule/SQLTranslatorRuleTest.java
@@ -20,7 +20,7 @@ package org.apache.shardingsphere.sqltranslator.rule;
 import org.apache.shardingsphere.infra.database.type.dialect.MySQLDatabaseType;
 import org.apache.shardingsphere.infra.database.type.dialect.PostgreSQLDatabaseType;
 import org.apache.shardingsphere.sqltranslator.api.config.SQLTranslatorRuleConfiguration;
-import org.apache.shardingsphere.sqltranslator.exception.UnsupportedTranslatedDatabaseException;
+import org.apache.shardingsphere.sqltranslator.exception.syntax.UnsupportedTranslatedDatabaseException;
 import org.junit.Test;
 
 import java.util.Locale;
